Simplified Application Analysis API
